+++ stdlib/Random/src/Random.jl
@@ -29,7 +29,7 @@ export rand!, randn!,
        randcycle, randcycle!,
        AbstractRNG, MersenneTwister, RandomDevice, TaskLocalRNG, Xoshiro
 
-public seed!, default_rng, Sampler, SamplerType, SamplerTrivial, SamplerSimple
+public seed!, default_rng, jump, jump!, Sampler, SamplerType, SamplerTrivial, SamplerSimple
 
 ## general definitions